Rename android app
