+++ apolloschurchapp/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.apolloschurchapp",
+    package = "com.bluebridgechurches.lcbc",
 )
 
 android_resource(
     name = "res",
-    package = "com.apolloschurchapp",
+    package = "com.bluebridgechurches.lcbc",
     res = "src/main/res",
 )
